+++ doc/changes.rst
@@ -5,7 +5,9 @@ desispec Change Log
 0.16.1 (unreleased)
 -------------------
 
-* No changes yet.
+* Fixed bug in :func:`desispec.parallel.dist_discrete` (PR `#446`_)
+
+.. _`#446`: https://github.com/desihub/desispec/pull/446
 
 0.16.0 (2017-09-29)
 -------------------